added assertion for continues
added assertion that checks that the continue length is divisible by the total length of the continue.
Had to create a separate xpath function for the total length bit as it wasnt possible before, took ages.
Also set some more stuff up for the positioning bit of the flatten XSL